Add  to relevent classes
