+++ geo/src/main/java/org/odk/collect/geo/GeoPointViewModel.kt
@@ -135,13 +135,4 @@ internal class LocationTrackerGeoPointViewModel(
     }
 }
 
-interface GeoPointViewModelFactory : ViewModelProvider.Factory
-
-internal sealed class GeoPointAccuracy {
-
-    abstract val value: Float
-
-    data class Improving(override val value: Float) : GeoPointAccuracy()
-    data class Poor(override val value: Float) : GeoPointAccuracy()
-    data class Unacceptable(override val value: Float) : GeoPointAccuracy()
-}
+internal interface GeoPointViewModelFactory : ViewModelProvider.Factory
